Removed Postgres Hard Dependency
Background: When using the tool, the postgres module was attempting to
be imported. However, if you are on a non-postgres system, then this
is not ideal. You don't want/need that dependency.

Solution: I've wrapped both the postgrs and mysql imports at the top
of the db.py file with try/except to catch ImportExceptions. For now they
will fail silently, until a better solution is discovered.