Lazily Init DB Connections
For some operations (list, gen-sql, gen-ref) we don't need to make
a connection to the DB. Additionally, we don't even need to provide
or use the DB credentials. This is paricularly true of the 'gen-sql'
command that the DBAs will be using on a regular basis. As such,
we want to make this optional (as it should be).

This is done by adding a new method 'init' to the DB classes that
only takes the config and sets a static var to say that the connection
has not been initialized. On the first attempt to 'execute' anything
against the DB, we initialize the connection.