AD-1539 - Script now operates scirpt-local
Before the tool would assume that it was one-directory below the
alters. This means that the tool was commonly added as a sub-module
to the main project. This is kinda crappy and means that you can
have multiple copies of the tool laying around. It also makes it hard
to add the command to your path because if you have multiple projects
it will be hard-wired to only work with one.

Now it works within whatever the current working directory is. This means
that one tool can be installed and added to the path. This also means that
the config can no longer live with the tool and will have to live with
the alters.