Added 'version'
You can now use '-v', '--version' or just 'version' with the
tool (no command) and the version of the tool will be shown.
The version is currently stored in a constant at the top of the
program. This works very similar to how '-h', '--help' and 'help'
commands/options are processed.