AD-1485 Directory Layout Cleanup
There was some old cruft to remove (Jenkins stuff) and the organization
of things was really getting messy. This just re-organizes everything and
removes the stuff that we don't need.

From this we can start breaking down `schema.py` to be a set of class-files
leaving the main file as just the bootstrapping file.