Updated 'up' command with '-u' flag
In the event that you want to run 'up' but you have some 'spotty' history,
you may not want to follow the default behavior of reconciling histories (
i.e. running the 'down' alters for anything that doesn't match the history
/ linearity of the files on disk). You can avoid this by passing in the '-u'
option. This does come at some risk though since you are going a little bit
outisde the paradigm of the tooling