Add require-env and skip-env metadata parameters.
-- require-env: foo, bar

In this case, if the 'env' in config.json is *not* one of 'foo' or 'bar'
then the alter is skipped.  If 'env' is not set in config.json, the
alter is skipped.

-- skip-env: baz, boy

In this case, if the 'env' in config.json is one of 'baz' or 'boy'
then the alter is skipped.  If 'env' is not set in config.json, the
alter is run.

Closes #3.