Fixed Skipped Envs
Anytime a value for skip or require-env was defined, it seemed to
be skipping the alter. Additionally for any alters that were skipped,
the up command would attempt to "unwind" the history back to the first
skipped alter, in an attempt to normalize the histories. The code has
now been corrected to know that it should not count those alters in the
history in the first place.