resolves #11 - base config + overrides
Introduces a base-configuration file, read from ~/.schema-tool
which can be overriden by project-configs (the config.json within
the project itself). This allows you to declutter your config
files for people managing multiple schema-repositories on a single
machine.

As an added bonus, the config loading was moved around so that you
can view the help file without having to load a config first, and thus
having to be within a project folder.