Removed `*` import from schema.py for commands
While it is super-convenient (and easier on the eyes) to import all the
commands with a `*` (I use them all anwyways), it also seems to make the
job of landscape.io harder. And I really want this tool to be effective
so I'll just make a super-long list that hurts your eyes so I can have this
bit of additional tooling. Trafeoffs 'eh.