Rewrite to GoLang - Initial Import
This is an in-progress commit of a rewrite of the schema tool to be in
Go. This is mainly a major cleanup of the tooling but also an effort to
simplify the distribution by being able to rely on single-binary
releases. The rewrite also intends to change the internal model to allow
for better exensibility and integration with other tooling.

This commit represents a cleanup of several initial commits and does not
consitute a working or complete rewrite of the tooling. It does however
include a compilable CLI, a working web-server, some initial tests,
scripts for licensing, and a usable Makefile along with some other
supporting code necessary for the port.